backend/fastapi/pyproject.toml
[project]
name = "autoadmin-fastapi-backend"
version = "1.0.0"
description = "Production-ready FastAPI backend for AutoAdmin AI agent orchestration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AutoAdmin Team", email = "team@autoadmin.com"}
]
keywords = ["fastapi", "ai", "agents", "langgraph", "openai", "async", "production"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # FastAPI and ASGI
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",

    # Pydantic for data validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database and ORM
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "databases[postgresql]>=0.8.0",

    # Redis for caching and Celery
    "redis>=5.0.0",
    "hiredis>=2.2.0",

    # Celery for background tasks
    "celery[redis]>=5.3.0",
    "flower>=2.0.0",

    # Firebase integration
    "firebase-admin>=6.2.0",
    "google-cloud-firestore>=2.16.0",
    "google-cloud-storage>=2.10.0",

    # OpenAI and AI services
    "openai>=1.6.0",
    "tiktoken>=0.5.0",

    # LangGraph and LangChain
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.26",

    # HTTP client for external services
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",

    # Authentication and security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",

    # File handling
    "python-magic>=0.4.27",
    "pillow>=10.1.0",
    "aiofiles>=23.2.0",

    # Vector operations
    "numpy>=1.25.0",
    "scipy>=1.11.0",

    # Monitoring and observability
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "python-json-logger>=2.0.7",

    # Rate limiting
    "slowapi>=0.1.9",

    # Date and time utilities
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",

    # Environment and configuration
    "python-dotenv>=1.0.0",

    # Data validation and serialization
    "email-validator>=2.1.0",
    "phonenumbers>=8.13.0",

    # Background processing
    "APScheduler>=3.10.0",

    # Webhook processing
    "cryptography>=41.0.0",
    "sigstore>=2.1.0",

    # Task queue monitoring
    "celery-redbeat>=2.0.0",

    # Performance monitoring
    "psutil>=5.9.0",
]

[dependency-groups]
dev = [
    # Development tools
    "black>=23.11.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",

    # Testing framework
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.25.0",  # For async testing

    # Code quality
    "pre-commit>=3.6.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",

    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",

    # Development database
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
monitoring = [
    "prometheus-fastapi-instrumentator>=6.1.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-exporter-prometheus>=1.12.0",
]

http-streaming = [
    "sse-starlette>=2.0.0",
]

all = [
    "autoadmin-fastapi-backend[monitoring,http-streaming]",
]

[project.scripts]
autoadmin-api = "main:app"
autoadmin-worker = "celery_app:worker"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "ARG001", # unused function argument
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG", "B"]

[tool.black]
target-version = ['py311']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "firebase_admin.*",
    "google.cloud.*",
    "openai.*",
    "langchain.*",
    "langgraph.*",
    "redis.*",
    "uvicorn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config --cov=app --cov-report=term-missing --cov-report=html --cov-report=xml"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "external: marks tests that require external services",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 80

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B601"]

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.11", "3.12"]